Drop brltty from the netinst CDs
brltty has started pulling in libicu, which has a size of 5MB.
An overhead of 5MB is just too big for netinst CD images for a
package that is only needed by a very small group of users and
that will normally also be available from a mirror (brltty gets
installed during finish-install).

For full CDs and DVDs the package will still get included as it's
listed in forcd1.