x86: update syslinux config manipulation for version 2.0
Adjust x86 boot scripts for new version (2.0) of framework in D-I for
creation of syslinux configuration files. Essentially the new version
is a backport of what was already implemented in debian-cd before the
release of Lenny and allows a significant simplification.

There are no significant changes in generated CD/DVD images.