Avoid introducing null bytes in module list in bug reports
Since the module taint string includes parentheses if is non-empty, I
originally set cut to use an empty output delimiter so that this
string would immediately follow the module name.  But currently this
results in a null byte between them, which unsurprisingly has led to
some slightly mangled bug reports.  For now, use sed to delete the
space.

svn path=/dists/trunk/linux-2.6/; revision=18512