rt2800usb: Re-enable powersaving by default
We've disabled powersaving for rt2800usb with a patch since 2.6.38-2,
as it resulted in very high latency then.

It looks like this has been fixed by commits
1c0bcf89d85cc97a0d9ce4cd909351a81fa4fdde ('rt2x00: Add autowake
support for USB hardware') and
5f0dd296a01c8173fcc05a8b262a1168ae90bc74 ('rt2x00: Implement
tx_frames_pending mac80211 callback function.').

Just comment-out the patch for now, in case we have to revert this.

svn path=/dists/sid/linux-2.6/; revision=19005