Use only one template syntax
Use the same template syntax and implementation for maintainer
scripts, translations, etc. as we do for the control files.  Define
the image-stem and initramfs variables to replace the old K and I
variables.

After this, debian/linux-* and debian/po/*.po are generated files (at
source preparation time) and should be ignored in svn.

Use debhelper to install the generated files at build time.  This also
results in a redundant dependency on debconf (which we already have in
Pre-Depends), but this seems harmless.

svn path=/dists/trunk/linux/; revision=20508