linux-image.postinst: Remove debconf warning about missing firmware
At this point, Debian users should know that they may need to install
firmware from non-free.  People using e.g. the r8169 driver may quite
reasonably choose not to install the associated firmware, either
because the driver doesn't actually request it for their chip or
because the driver can still work without it.

One thing we lose by doing this is a reminder that a firmware package
might also need to be upgraded, as a driver requires a newer version
of the firmware that has a different name.  As an alternative, we
could compare the firmware file lists for old and new modules and only
warn about newly listed files that are missing.  However, that would
also result in incorrect warnings for e.g. r8169 users, as that driver
may request a different file for each of the many chips it supports.

svn path=/dists/trunk/linux/; revision=20511