Avoid building with libunwind8-dev
The transition from libunwind soversion 7 to 8 is blocked and will
likely remain blocked for some time because libunwind7 is linked into
all ia64 binaries.

For amd64 and i386, downgrade the build-dependency to libunwind7-dev.

For architectures that are supported by libunwind8-dev, remove the
build-dependency and ensure we don't accidentally link with libunwind8
by setting NO_LIBUNWIND=1 when building perf.

svn path=/dists/sid/linux-tools/; revision=20755