aufs: Update to aufs3.x-rcN-20131104
mmap() holds reference to the union file as well as the branch file.
This behaviour requires an extra patch and used to be conditional on
CONFIG_AUFS_PROC_MAP.  It is now unconditional because otherwise the
aufs inode number may be reused and this can break userland file
caching.  So add the extra patch and drop our patch to disable
CONFIG_AUFS_PROC_MAP.

svn path=/dists/trunk/linux/; revision=20778