Bypass the new and buggy tools/perf/Makefile in favour of Makefile.perf
Somehow this new wrapper manages to reduce the priority of our
'prefix' setting.  That interacts with the insane way prefix is
defaulted in config/Makefile so that the prefix is wrong at build time
but right at install time (resulting in a rebuild then!).

Also the wrapper replaces our -j option and we would have to set JOBS
instead.

The perf makefiles never get better, just different...

svn path=/dists/trunk/linux-tools/; revision=20994