[powerpc/powerpc64] Re-enable JUMP_LABEL
CONFIG_JUMP_LABEL only has an effect if the compiler properly supports
"asm goto".  That broke for powerpc 64-bit code in some version of
gcc-4.8, and when it was fixed that changed ABI so we had to disable
the config symbol temporarily.

Make sure we get the fixed gcc-4.8 by using a versioned build-dep.

svn path=/dists/sid/linux/; revision=21570