udeb: Fix typo in dependencies of speakup-modules
We currently FTBFS on mips64el:

    cannot read /tmp/linux/linux-3.16.2/debian/installer/mips64el/modules/mips64el-sb1-bcm91250a/mips64el-sb1-bcm91250a
    ...
    cannot read /tmp/linux/linux-3.16.2/debian/installer/mips64el/modules/mips64el-loongson-3/mips64el-loongson-3
    ...
    find: `debian/speakup-modules-3.16-1-loongson-3-di': No such file or directory
    speakup-modules-3.16-1-loongson-3-di will be empty

The 'cannot read' warnings are due to a bug in kernel-wedge
copy-modules (fixed in version 2.90) that affects module list
directories that are symlinked.  Instead of getting a list of files
and symlinks in the directory, it only finds the directory-symlink.

The bug was generally harmless because copy-modules also gets a
list of inter-package dependencies (without relying on the first
list), and this includes all package names.  Except that
speakup-modules has 'Depends: kernel-image!', and as there is
no package definition for 'kernel-image!' this dependency is
not included in the second list and copy-modules does not see
'speakup-modules'.

mips64el seems to be the only architecture which has a symlink to a
module list directory including speakup-modules.

Delete the '!'.

svn path=/dists/sid/linux/; revision=21834