Major refactor for better control of choice of EFI and BIOS boot
Split the code up and use BOOT_EFI and BOOT_BIOS to control which
goes where. We now support:

 * BOOT_BIOS=1 BOOT_EFI=0 for BIOS boot only via an ElTorito boot
   record (like debian-cd used to be before EFI support was added)

 * BOOT_BIOS=1 BOOT_EFI=2 for standard BIOS boot as
   primary ElTorito boot record and EFI as an alternate ElTorito boot
   record (common case, just like we've been doing for amd64 in
   debian-cd for a while)

 * BOOT_BIOS=0 and BOOT_EFI=1 for *only* EFI boot as the primary
   ElTorito boot record (new case, might be useful for some
   Macs *maybe*)

Set BOOT_METHODS in the debian-cd environment to determine which of
these cases is desired in a given build.

Also removed the support for the "old" syslinux packaging layout, it's
not been around for a while now.