Improve the sort_deps fix for too-large packages
Make sure that the package we're told about exists before we look for
its size - a number of the dependencies and recommends may not exist
and this will cause errors for us.

Also list the configured max_pkg_size at the top of the script, to
help with debugging later.