debian/control: Correct build-dependency on xz-utils, needed on all arches
Currently we have xz-utils in build-depends-indep as we explicitly
use xz to compress files in linux-source.  We also have a build-
dependency on lzma (now a virtual package provided by xz-utils),
but only on armel which was the first architecture for which we
set CONFIG_KERNEL_XZ=y.

However, since 3.7 we've set CONFIG_KERNEL_XZ for all architectures so
xz-utils is needed everywhere.  Currently we get away with this
because dpkg-dev depends on xz-utils, but this could change in future.
Build-depend on xz-utils without qualification.