Fix size calculation for EFI image
CD creation fails with "disk full", because of rounding errors.

The minimum size of data-allocation is a 'cluster', which is a
power-of-two multiple of sectors.
mkfs.msdos chooses minimum FAT and cluster size for image: FAT12<16G
mkfs.msdos uses block-count where a block is 1024 B, but must be rounded
up to track_size with 32 sectors.

Round up each file to clusters before summing.
A sub-directory entry needs 1 cluster minimum.