Add support for the nopython build profile, disabling the linux-perf package
Currently we build-depend on the native python (via asciidoc), and on
the host python (via python-dev).  As these are not coinstallable it
is impossible to perform a complete cross-build.  Until that's resolved,
this will allow cross-building of most of the package with the
combination of the 'cross' and 'nopython' profiles.

(This also sidesteps the issue of perf wanting a multilib compiler.)