[powerpc,powerpcspe,ppc64] linux-image: Suppress automatic dbgsym packages
Currently on powerpc, powerpcspe and ppc64 we get an automatic dbgsym
package with symbols for the bootwrapper tools (addnote, hack-coff,
mktree).  We should either put them in linux-image-*-dbgsym or
nowhere.  For now, opt for nowhere.

Move the dh_strip invocation from the install-base rule to the
install-image_... rule.  None of the other packages using install-base
should contain any executables.