[arm64] video: Change FB from module to built-in
Currently we don't explicitly set CONFIG_FB for arm64, and the only
reason it's enabled as a module is that the DRM drivers (which are
built as modules) select it.  We *do* explicitly set CONFIG_FB=y on
almost all other architectures and flavours.

Simplify this by setting CONFIG_FB=y at the top level and overriding
where needed (not arm64).