linux-doc: Copy source to build directory
This is necessary to avoid creating files in source directory, which
makes the linux-source package unclean (and unreproducible).

Firstly, Python creates bytecode files alongside the module sources.
We can and do exclude those, though.

Secondly, starting with 4.10, Documentation/media/Makefile converts
from dot to SVG and from SVG to PDF in the source directory.  These
can't easily be excluded, as SVG is also used as a source format.