Install copyright file when the nodoc profile is used
Revert to running dh_installdocs unconditionally, although that
currently installs more than we want (which is permitted by policy).

When we upgrade to debhelper compat level 11, dh_installdocs will
become sensitive to the profile and will install only the copyright
file in this case.  But we shouldn't do that until development of
this level is complete and supported in stretch-backports.