debian/lib/python/debian_linux/debian.py: Parse bottom lines of changelog entries
Changelog doesn't currently provide access to the maintainer
name/address and date.  We need this when updating the signed template
changelog.

While we're at it, make sure we don't ignore any important lines.
Anything beginning with exactly zero or one spaces is a top or bottom
line, respectively; anything else is internal text we can ignore.