ext4: Fix duplicate softdep fields in module info
We no longer need to add the crypto-aes or crypto-ecb dependencies
because ext4 uses the common encryption code in fscrypto (and has a
regular symbol dependency on it).

Since upstream added a softdep on "crc32c", we don't actually need to
change ext4 at all now.  But let's replace it with "crypto-crc32c",
since that's the module alias the crypto subsystem will actually
request and is consistent with the softdep we add to other
filesystems.