debian/lib/python/debian_linux/debian.py: Fix Version self-test failures
We didn't check for invalid characters in the upstream or revision
parts.  Tighten the regexps for those parts.

Also, failure to match the epoch or revision groups caused
those parts to be treated as part of the upstream version.
Split out the three parts and match them against separate
regexps.