debian/lib/python/debian_linux/debian.py: Accept arbitrary revision suffixes
Closes: #898087

We always used to match arbitrary revision suffixes, but in commit
6a4f7206e254 ".../debian.py: Fix binNMU revision parsing" I excluded
'+' from the revision_other to ensure that it wouldn't match a
binNMU suffix (which is matched separately).

Stop excluding '+' from the revision_other group, and change the
repetition to non-greedy so that any binNMU suffix will still match
the following group.