debian/control: Remove "cross" from profiles for build-dep on libssl-dev
Ideally we should not require any special profile to be enabled for a
cross-build.  (Unfortunately we still will because cross-compiler
and native compiler packages are named differently.)

For a native build, the build-dependency on libssl-dev is redundant
with the previous build-dependency on libssl-dev:native, so this
shouldn't make any practical difference.