debian/rules: Checksum only the source name and version from debian/changelog
Checksumming the whole of debian/changelog when deciding whether to
run gencontrol.py results in (a) frequent changes to control.md5sum
and (b) the need to invoke various targets twice during development.

I originally made this change to address (a), which would be an
annoyance if and when we start using dgit.  However, fixing (b) is a
nice benefit regardless of whether we do that.