[powerpc,powerpcspe,ppc64] linux-config: Eliminate config.*_bootwrapper.gz files
On powerpc architectures that may use a bootwrapper, we create a
temporary build_<arch>_<featureset>_<flavour>_bootwrapper directory
for each kernel configuration to hold the related tools which we won't
install for real (because they are always native).

This directory is then matched by the wildcard used in building
linux-config, causing linux-config packages to contain spurious
(empty) kconfig files based on these directories in addition to the
real kconfig files.

Rename the temporary directory to avoid matching that wildcard.