Munge the EFI boot support code to make Secure Boot work
We don't need to do much here, just pass on what the d-i build already
set up for us. Previously, we assumed too much knowledge about the
contents of the efi.img we were given. Now simply copy all of it.

(This is only really necessary due to us wanting to merge the efi.img
 for multiple architectures. This is likely to get harder with
 clashing files... :-/)