[ppc64el] Disable PCMCIA (fixes FTBFS)
It appears to be technically possible to use PCMCIA cards on POWER8/9
systems through a PCI Express to PCI adapter and a PCI to
PCMCIA/CardBus adapter.  But I can't believe anyone would want to.

So rather than adding a pcmcia-modules package or excluding the
drivers from udebs, disable PCMCIA altogether.